streamline hang planners
